+++ example_dartmapper/lib/example.g.dart
@@ -8,10 +8,10 @@ part of 'example.dart';
 
 Task _$TaskFromJson(Map<String, dynamic> json) {
   return Task(
-    id: json['id'] as String,
-    name: json['name'] as String,
-    avatar: json['avatar'] as String,
-    createdAt: json['createdAt'] as String,
+    id: json['id'] as String?,
+    name: json['name'] as String?,
+    avatar: json['avatar'] as String?,
+    createdAt: json['createdAt'] as String?,
   );
 }
 
@@ -24,46 +24,40 @@ Map<String, dynamic> _$TaskToJson(Task instance) => <String, dynamic>{
 
 TaskQuery _$TaskQueryFromJson(Map<String, dynamic> json) {
   return TaskQuery(
-    (json['statuses'] as List)
-        ?.map((e) => _$enumDecodeNullable(_$StatusEnumMap, e))
-        ?.toList(),
+    (json['statuses'] as List<dynamic>)
+        .map((e) => _$enumDecode(_$StatusEnumMap, e))
+        .toList(),
   );
 }
 
 Map<String, dynamic> _$TaskQueryToJson(TaskQuery instance) => <String, dynamic>{
-      'statuses': instance.statuses?.map((e) => _$StatusEnumMap[e])?.toList(),
+      'statuses': instance.statuses.map((e) => _$StatusEnumMap[e]).toList(),
     };
 
-T _$enumDecode<T>(
-  Map<T, dynamic> enumValues,
-  dynamic source, {
-  T unknownValue,
+K _$enumDecode<K, V>(
+  Map<K, V> enumValues,
+  Object? source, {
+  K? unknownValue,
 }) {
   if (source == null) {
-    throw ArgumentError('A value must be provided. Supported values: '
-        '${enumValues.values.join(', ')}');
+    throw ArgumentError(
+      'A value must be provided. Supported values: '
+      '${enumValues.values.join(', ')}',
+    );
   }
 
-  final value = enumValues.entries
-      .singleWhere((e) => e.value == source, orElse: () => null)
-      ?.key;
-
-  if (value == null && unknownValue == null) {
-    throw ArgumentError('`$source` is not one of the supported values: '
-        '${enumValues.values.join(', ')}');
-  }
-  return value ?? unknownValue;
-}
-
-T _$enumDecodeNullable<T>(
-  Map<T, dynamic> enumValues,
-  dynamic source, {
-  T unknownValue,
-}) {
-  if (source == null) {
-    return null;
-  }
-  return _$enumDecode<T>(enumValues, source, unknownValue: unknownValue);
+  return enumValues.entries.singleWhere(
+    (e) => e.value == source,
+    orElse: () {
+      if (unknownValue == null) {
+        throw ArgumentError(
+          '`$source` is not one of the supported values: '
+          '${enumValues.values.join(', ')}',
+        );
+      }
+      return MapEntry(unknownValue, enumValues.values.first);
+    },
+  ).key;
 }
 
 const _$StatusEnumMap = {
@@ -75,18 +69,15 @@ const _$StatusEnumMap = {
 TaskGroup _$TaskGroupFromJson(Map<String, dynamic> json) {
   return TaskGroup(
     date: json['date'] == null ? null : DateTime.parse(json['date'] as String),
-    todos: (json['todos'] as List)
-        ?.map(
-            (e) => e == null ? null : Task.fromJson(e as Map<String, dynamic>))
-        ?.toList(),
-    completed: (json['completed'] as List)
-        ?.map(
-            (e) => e == null ? null : Task.fromJson(e as Map<String, dynamic>))
-        ?.toList(),
-    inProgress: (json['inProgress'] as List)
-        ?.map(
-            (e) => e == null ? null : Task.fromJson(e as Map<String, dynamic>))
-        ?.toList(),
+    todos: (json['todos'] as List<dynamic>?)
+        ?.map((e) => Task.fromJson(e as Map<String, dynamic>))
+        .toList(),
+    completed: (json['completed'] as List<dynamic>?)
+        ?.map((e) => Task.fromJson(e as Map<String, dynamic>))
+        .toList(),
+    inProgress: (json['inProgress'] as List<dynamic>?)
+        ?.map((e) => Task.fromJson(e as Map<String, dynamic>))
+        .toList(),
   );
 }
 
@@ -103,28 +94,25 @@ Map<String, dynamic> _$TaskGroupToJson(TaskGroup instance) => <String, dynamic>{
 
 class _RestClient implements RestClient {
   _RestClient(this._dio, {this.baseUrl}) {
-    ArgumentError.checkNotNull(_dio, '_dio');
     baseUrl ??= 'https://5d42a6e2bc64f90014a56ca0.mockapi.io/api/v1/';
   }
 
   final Dio _dio;
 
-  String baseUrl;
+  String? baseUrl;
 
   @override
   Future<List<String>> getTags() async {
     const _extra = <String, dynamic>{};
     final queryParameters = <String, dynamic>{};
     final _data = <String, dynamic>{};
-    final _result = await _dio.request<List<dynamic>>('/tags',
-        queryParameters: queryParameters,
-        options: RequestOptions(
-            method: 'GET',
-            headers: <String, dynamic>{},
-            extra: _extra,
-            baseUrl: baseUrl),
-        data: _data);
-    final value = _result.data.cast<String>();
+    final _result = await _dio.fetch<List<dynamic>>(
+        _setStreamType<List<String>>(
+            Options(method: 'GET', headers: <String, dynamic>{}, extra: _extra)
+                .compose(_dio.options, '/tags',
+                    queryParameters: queryParameters, data: _data)
+                .copyWith(baseUrl: baseUrl ?? _dio.options.baseUrl)));
+    final value = _result.data!.cast<String>();
     return value;
   }
 
@@ -133,15 +121,13 @@ class _RestClient implements RestClient {
     const _extra = <String, dynamic>{};
     final queryParameters = <String, dynamic>{};
     final _data = <String, dynamic>{};
-    final _result = await _dio.request<List<dynamic>>('/tags',
-        queryParameters: queryParameters,
-        options: RequestOptions(
-            method: 'GET',
-            headers: <String, dynamic>{},
-            extra: _extra,
-            baseUrl: baseUrl),
-        data: _data);
-    final value = _result.data.cast<String>();
+    final _result = await _dio.fetch<List<dynamic>>(
+        _setStreamType<List<String>>(
+            Options(method: 'GET', headers: <String, dynamic>{}, extra: _extra)
+                .compose(_dio.options, '/tags',
+                    queryParameters: queryParameters, data: _data)
+                .copyWith(baseUrl: baseUrl ?? _dio.options.baseUrl)));
+    final value = _result.data!.cast<String>();
     yield value;
   }
 
@@ -150,15 +136,12 @@ class _RestClient implements RestClient {
     const _extra = <String, dynamic>{};
     final queryParameters = <String, dynamic>{};
     final _data = <String, dynamic>{};
-    final _result = await _dio.request<List<dynamic>>('/tasks',
-        queryParameters: queryParameters,
-        options: RequestOptions(
-            method: 'GET',
-            headers: <String, dynamic>{},
-            extra: _extra,
-            baseUrl: baseUrl),
-        data: _data);
-    var value = _result.data
+    final _result = await _dio.fetch<List<dynamic>>(_setStreamType<List<Task>>(
+        Options(method: 'GET', headers: <String, dynamic>{}, extra: _extra)
+            .compose(_dio.options, '/tasks',
+                queryParameters: queryParameters, data: _data)
+            .copyWith(baseUrl: baseUrl ?? _dio.options.baseUrl)));
+    var value = _result.data!
         .map((dynamic i) => Task.fromJson(i as Map<String, dynamic>))
         .toList();
     return value;
@@ -166,104 +149,78 @@ class _RestClient implements RestClient {
 
   @override
   Future<Task> getTask(id) async {
-    ArgumentError.checkNotNull(id, 'id');
     const _extra = <String, dynamic>{};
     final queryParameters = <String, dynamic>{};
     final _data = <String, dynamic>{};
-    final _result = await _dio.request<Map<String, dynamic>>('/tasks/$id',
-        queryParameters: queryParameters,
-        options: RequestOptions(
-            method: 'GET',
-            headers: <String, dynamic>{},
-            extra: _extra,
-            baseUrl: baseUrl),
-        data: _data);
-    final value = Task.fromJson(_result.data);
+    final _result = await _dio.fetch<Map<String, dynamic>>(_setStreamType<Task>(
+        Options(method: 'GET', headers: <String, dynamic>{}, extra: _extra)
+            .compose(_dio.options, '/tasks/$id',
+                queryParameters: queryParameters, data: _data)
+            .copyWith(baseUrl: baseUrl ?? _dio.options.baseUrl)));
+    final value = Task.fromJson(_result.data!);
     return value;
   }
 
   @override
   Future<Task> updateTaskPart(id, map) async {
-    ArgumentError.checkNotNull(id, 'id');
-    ArgumentError.checkNotNull(map, 'map');
     const _extra = <String, dynamic>{};
     final queryParameters = <String, dynamic>{};
     final _data = <String, dynamic>{};
-    _data.addAll(map ?? <String, dynamic>{});
-    _data.removeWhere((k, v) => v == null);
-    final _result = await _dio.request<Map<String, dynamic>>('/tasks/$id',
-        queryParameters: queryParameters,
-        options: RequestOptions(
-            method: 'PATCH',
-            headers: <String, dynamic>{},
-            extra: _extra,
-            baseUrl: baseUrl),
-        data: _data);
-    final value = Task.fromJson(_result.data);
+    _data.addAll(map);
+    final _result = await _dio.fetch<Map<String, dynamic>>(_setStreamType<Task>(
+        Options(method: 'PATCH', headers: <String, dynamic>{}, extra: _extra)
+            .compose(_dio.options, '/tasks/$id',
+                queryParameters: queryParameters, data: _data)
+            .copyWith(baseUrl: baseUrl ?? _dio.options.baseUrl)));
+    final value = Task.fromJson(_result.data!);
     return value;
   }
 
   @override
   Future<Task> updateTask(id, task) async {
-    ArgumentError.checkNotNull(id, 'id');
-    ArgumentError.checkNotNull(task, 'task');
     const _extra = <String, dynamic>{};
     final queryParameters = <String, dynamic>{};
     final _data = <String, dynamic>{};
-    _data.addAll(task?.toJson() ?? <String, dynamic>{});
-    _data.removeWhere((k, v) => v == null);
-    final _result = await _dio.request<Map<String, dynamic>>('/tasks/$id',
-        queryParameters: queryParameters,
-        options: RequestOptions(
-            method: 'PUT',
-            headers: <String, dynamic>{},
-            extra: _extra,
-            baseUrl: baseUrl),
-        data: _data);
-    final value = Task.fromJson(_result.data);
+    _data.addAll(task.toJson());
+    final _result = await _dio.fetch<Map<String, dynamic>>(_setStreamType<Task>(
+        Options(method: 'PUT', headers: <String, dynamic>{}, extra: _extra)
+            .compose(_dio.options, '/tasks/$id',
+                queryParameters: queryParameters, data: _data)
+            .copyWith(baseUrl: baseUrl ?? _dio.options.baseUrl)));
+    final value = Task.fromJson(_result.data!);
     return value;
   }
 
   @override
   Future<void> deleteTask(id) async {
-    ArgumentError.checkNotNull(id, 'id');
     const _extra = <String, dynamic>{};
     final queryParameters = <String, dynamic>{};
     final _data = <String, dynamic>{};
-    await _dio.request<void>('/tasks/$id',
-        queryParameters: queryParameters,
-        options: RequestOptions(
-            method: 'DELETE',
-            headers: <String, dynamic>{},
-            extra: _extra,
-            baseUrl: baseUrl),
-        data: _data);
+    await _dio.fetch<void>(_setStreamType<void>(
+        Options(method: 'DELETE', headers: <String, dynamic>{}, extra: _extra)
+            .compose(_dio.options, '/tasks/$id',
+                queryParameters: queryParameters, data: _data)
+            .copyWith(baseUrl: baseUrl ?? _dio.options.baseUrl)));
     return null;
   }
 
   @override
   Future<Task> createTask(task) async {
-    ArgumentError.checkNotNull(task, 'task');
     const _extra = <String, dynamic>{};
     final queryParameters = <String, dynamic>{};
     final _data = <String, dynamic>{};
-    _data.addAll(task?.toJson() ?? <String, dynamic>{});
-    _data.removeWhere((k, v) => v == null);
-    final _result = await _dio.request<Map<String, dynamic>>('/tasks',
-        queryParameters: queryParameters,
-        options: RequestOptions(
-            method: 'POST',
-            headers: <String, dynamic>{},
-            extra: _extra,
-            baseUrl: baseUrl),
-        data: _data);
-    final value = Task.fromJson(_result.data);
+    _data.addAll(task.toJson());
+    final _result = await _dio.fetch<Map<String, dynamic>>(_setStreamType<Task>(
+        Options(method: 'POST', headers: <String, dynamic>{}, extra: _extra)
+            .compose(_dio.options, '/tasks',
+                queryParameters: queryParameters, data: _data)
+            .copyWith(baseUrl: baseUrl ?? _dio.options.baseUrl)));
+    final value = Task.fromJson(_result.data!);
     return value;
   }
 
   @override
   Future<void> createNewTaskFromFile(file) async {
-    ArgumentError.checkNotNull(file, 'file');
     const _extra = <String, dynamic>{};
     final queryParameters = <String, dynamic>{};
     final _data = FormData();
@@ -271,14 +228,11 @@ class _RestClient implements RestClient {
         'file',
         MultipartFile.fromFileSync(file.path,
             filename: file.path.split(Platform.pathSeparator).last)));
-    await _dio.request<void>('http://httpbin.org/post',
-        queryParameters: queryParameters,
-        options: RequestOptions(
-            method: 'POST',
-            headers: <String, dynamic>{},
-            extra: _extra,
-            baseUrl: baseUrl),
-        data: _data);
+    await _dio.fetch<void>(_setStreamType<void>(
+        Options(method: 'POST', headers: <String, dynamic>{}, extra: _extra)
+            .compose(_dio.options, 'http://httpbin.org/post',
+                queryParameters: queryParameters, data: _data)
+            .copyWith(baseUrl: baseUrl ?? _dio.options.baseUrl)));
     return null;
   }
 
@@ -287,38 +241,33 @@ class _RestClient implements RestClient {
     const _extra = <String, dynamic>{};
     final queryParameters = <String, dynamic>{};
     final _data = <String, dynamic>{};
-    final _result = await _dio.request<List<dynamic>>(
-        'http://httpbin.org/image/jpeg',
-        queryParameters: queryParameters,
-        options: RequestOptions(
-            method: 'GET',
-            headers: <String, dynamic>{r'accept': 'image/jpeg'},
-            extra: _extra,
-            baseUrl: baseUrl,
-            responseType: ResponseType.bytes),
-        data: _data);
-    final value = _result.data.cast<int>();
+    final _result = await _dio.fetch<List<dynamic>>(_setStreamType<List<int>>(
+        Options(
+                method: 'GET',
+                headers: <String, dynamic>{r'accept': 'image/jpeg'},
+                extra: _extra,
+                responseType: ResponseType.bytes)
+            .compose(_dio.options, 'http://httpbin.org/image/jpeg',
+                queryParameters: queryParameters, data: _data)
+            .copyWith(baseUrl: baseUrl ?? _dio.options.baseUrl)));
+    final value = _result.data!.cast<int>();
     return value;
   }
 
   @override
-  Future<String> postUrlEncodedFormData(hello, {gg}) async {
-    ArgumentError.checkNotNull(hello, 'hello');
+  Future<String> postUrlEncodedFormData(hello, {required gg}) async {
     const _extra = <String, dynamic>{};
     final queryParameters = <String, dynamic>{};
-    queryParameters.removeWhere((k, v) => v == null);
     final _data = {'hello': hello, 'gg': gg};
-    _data.removeWhere((k, v) => v == null);
-    final _result = await _dio.request<String>('http://httpbin.org/post',
-        queryParameters: queryParameters,
-        options: RequestOptions(
+    final _result = await _dio.fetch<String>(_setStreamType<String>(Options(
             method: 'POST',
             headers: <String, dynamic>{},
             extra: _extra,
-            contentType: 'application/x-www-form-urlencoded',
-            baseUrl: baseUrl),
-        data: _data);
-    final value = _result.data;
+            contentType: 'application/x-www-form-urlencoded')
+        .compose(_dio.options, 'http://httpbin.org/post',
+            queryParameters: queryParameters, data: _data)
+        .copyWith(baseUrl: baseUrl ?? _dio.options.baseUrl)));
+    final value = _result.data!;
     return value;
   }
 
@@ -327,15 +276,12 @@ class _RestClient implements RestClient {
     const _extra = <String, dynamic>{};
     final queryParameters = <String, dynamic>{};
     final _data = <String, dynamic>{};
-    final _result = await _dio.request<String>('/',
-        queryParameters: queryParameters,
-        options: RequestOptions(
-            method: 'HEAD',
-            headers: <String, dynamic>{},
-            extra: _extra,
-            baseUrl: baseUrl),
-        data: _data);
-    final value = _result.data;
+    final _result = await _dio.fetch<String>(_setStreamType<String>(
+        Options(method: 'HEAD', headers: <String, dynamic>{}, extra: _extra)
+            .compose(_dio.options, '/',
+                queryParameters: queryParameters, data: _data)
+            .copyWith(baseUrl: baseUrl ?? _dio.options.baseUrl)));
+    final value = _result.data!;
     return value;
   }
 
@@ -344,15 +290,12 @@ class _RestClient implements RestClient {
     const _extra = <String, dynamic>{};
     final queryParameters = <String, dynamic>{};
     final _data = <String, dynamic>{};
-    final _result = await _dio.request('/',
-        queryParameters: queryParameters,
-        options: RequestOptions(
-            method: 'HEAD',
-            headers: <String, dynamic>{},
-            extra: _extra,
-            baseUrl: baseUrl),
-        data: _data);
-    final value = _result.data;
+    final _result = await _dio.fetch(_setStreamType<dynamic>(
+        Options(method: 'HEAD', headers: <String, dynamic>{}, extra: _extra)
+            .compose(_dio.options, '/',
+                queryParameters: queryParameters, data: _data)
+            .copyWith(baseUrl: baseUrl ?? _dio.options.baseUrl)));
+    final value = _result.data!;
     return value;
   }
 
@@ -361,15 +304,12 @@ class _RestClient implements RestClient {
     const _extra = <String, dynamic>{};
     final queryParameters = <String, dynamic>{};
     final _data = <String, dynamic>{};
-    final _result = await _dio.request('/',
-        queryParameters: queryParameters,
-        options: RequestOptions(
-            method: 'HEAD',
-            headers: <String, dynamic>{},
-            extra: _extra,
-            baseUrl: baseUrl),
-        data: _data);
-    final value = _result.data;
+    final _result = await _dio.fetch(_setStreamType<HttpResponse<dynamic>>(
+        Options(method: 'HEAD', headers: <String, dynamic>{}, extra: _extra)
+            .compose(_dio.options, '/',
+                queryParameters: queryParameters, data: _data)
+            .copyWith(baseUrl: baseUrl ?? _dio.options.baseUrl)));
+    final value = _result.data!;
     final httpResponse = HttpResponse(value, _result);
     return httpResponse;
   }
@@ -379,15 +319,13 @@ class _RestClient implements RestClient {
     const _extra = <String, dynamic>{};
     final queryParameters = <String, dynamic>{};
     final _data = <String, dynamic>{};
-    final _result = await _dio.request<List<dynamic>>('/task/group',
-        queryParameters: queryParameters,
-        options: RequestOptions(
-            method: 'GET',
-            headers: <String, dynamic>{},
-            extra: _extra,
-            baseUrl: baseUrl),
-        data: _data);
-    var value = _result.data
+    final _result = await _dio.fetch<List<dynamic>>(
+        _setStreamType<List<TaskGroup>>(
+            Options(method: 'GET', headers: <String, dynamic>{}, extra: _extra)
+                .compose(_dio.options, '/task/group',
+                    queryParameters: queryParameters, data: _data)
+                .copyWith(baseUrl: baseUrl ?? _dio.options.baseUrl)));
+    var value = _result.data!
         .map((dynamic i) => TaskGroup.fromJson(i as Map<String, dynamic>))
         .toList();
     return value;
@@ -398,15 +336,13 @@ class _RestClient implements RestClient {
     const _extra = <String, dynamic>{};
     final queryParameters = <String, dynamic>{};
     final _data = <String, dynamic>{};
-    final _result = await _dio.request<List<dynamic>>('/task',
-        queryParameters: queryParameters,
-        options: RequestOptions(
-            method: 'GET',
-            headers: <String, dynamic>{},
-            extra: _extra,
-            baseUrl: baseUrl),
-        data: _data);
-    var value = _result.data
+    final _result = await _dio.fetch<List<dynamic>>(
+        _setStreamType<HttpResponse<List<Task>>>(
+            Options(method: 'GET', headers: <String, dynamic>{}, extra: _extra)
+                .compose(_dio.options, '/task',
+                    queryParameters: queryParameters, data: _data)
+                .copyWith(baseUrl: baseUrl ?? _dio.options.baseUrl)));
+    var value = _result.data!
         .map((dynamic i) => Task.fromJson(i as Map<String, dynamic>))
         .toList();
     final httpResponse = HttpResponse(value, _result);
@@ -415,60 +351,46 @@ class _RestClient implements RestClient {
 
   @override
   Future<HttpResponse<void>> deleteTaskWithResponse(id) async {
-    ArgumentError.checkNotNull(id, 'id');
     const _extra = <String, dynamic>{};
     final queryParameters = <String, dynamic>{};
     final _data = <String, dynamic>{};
-    final _result = await _dio.request<void>('/tasks/$id',
-        queryParameters: queryParameters,
-        options: RequestOptions(
-            method: 'DELETE',
-            headers: <String, dynamic>{},
-            extra: _extra,
-            baseUrl: baseUrl),
-        data: _data);
+    final _result = await _dio.fetch<void>(_setStreamType<HttpResponse<void>>(
+        Options(method: 'DELETE', headers: <String, dynamic>{}, extra: _extra)
+            .compose(_dio.options, '/tasks/$id',
+                queryParameters: queryParameters, data: _data)
+            .copyWith(baseUrl: baseUrl ?? _dio.options.baseUrl)));
     final httpResponse = HttpResponse(null, _result);
     return httpResponse;
   }
 
   @override
-  Future<String> postFormData(task, {file}) async {
-    ArgumentError.checkNotNull(task, 'task');
+  Future<String> postFormData(task, {required file}) async {
     const _extra = <String, dynamic>{};
     final queryParameters = <String, dynamic>{};
-    queryParameters.removeWhere((k, v) => v == null);
     final _data = FormData();
-    _data.fields.add(MapEntry('task', jsonEncode(task ?? <String, dynamic>{})));
-    if (file != null) {
-      _data.files.add(MapEntry(
-          'file',
-          MultipartFile.fromFileSync(file.path,
-              filename: file.path.split(Platform.pathSeparator).last)));
-    }
-    final _result = await _dio.request<String>('/post',
-        queryParameters: queryParameters,
-        options: RequestOptions(
-            method: 'POST',
-            headers: <String, dynamic>{},
-            extra: _extra,
-            baseUrl: baseUrl),
-        data: _data);
-    final value = _result.data;
+    _data.fields.add(MapEntry('task', jsonEncode(task)));
+    _data.files.add(MapEntry(
+        'file',
+        MultipartFile.fromFileSync(file.path,
+            filename: file.path.split(Platform.pathSeparator).last)));
+    final _result = await _dio.fetch<String>(_setStreamType<String>(
+        Options(method: 'POST', headers: <String, dynamic>{}, extra: _extra)
+            .compose(_dio.options, '/post',
+                queryParameters: queryParameters, data: _data)
+            .copyWith(baseUrl: baseUrl ?? _dio.options.baseUrl)));
+    final value = _result.data!;
     return value;
   }
 
   @override
   Future<String> postFormData2(task, tags, file) async {
-    ArgumentError.checkNotNull(task, 'task');
-    ArgumentError.checkNotNull(tags, 'tags');
-    ArgumentError.checkNotNull(file, 'file');
     const _extra = <String, dynamic>{};
     final queryParameters = <String, dynamic>{};
     final _data = FormData();
-    task?.forEach((i) {
+    task.forEach((i) {
       _data.fields.add(MapEntry('task', jsonEncode(i)));
     });
-    tags?.forEach((i) {
+    tags.forEach((i) {
       _data.fields.add(MapEntry('tags', i));
     });
     _data.files.add(MapEntry(
@@ -476,56 +398,46 @@ class _RestClient implements RestClient {
         MultipartFile.fromFileSync(file.path,
             filename: file.path.split(Platform.pathSeparator).last,
             contentType: MediaType.parse('application/json'))));
-    final _result = await _dio.request<String>('/post',
-        queryParameters: queryParameters,
-        options: RequestOptions(
-            method: 'POST',
-            headers: <String, dynamic>{},
-            extra: _extra,
-            baseUrl: baseUrl),
-        data: _data);
-    final value = _result.data;
+    final _result = await _dio.fetch<String>(_setStreamType<String>(
+        Options(method: 'POST', headers: <String, dynamic>{}, extra: _extra)
+            .compose(_dio.options, '/post',
+                queryParameters: queryParameters, data: _data)
+            .copyWith(baseUrl: baseUrl ?? _dio.options.baseUrl)));
+    final value = _result.data!;
     return value;
   }
 
   @override
-  Future<String> postFormData3({files, file}) async {
+  Future<String> postFormData3({required files, required file}) async {
     const _extra = <String, dynamic>{};
     final queryParameters = <String, dynamic>{};
-    queryParameters.removeWhere((k, v) => v == null);
     final _data = FormData();
-    _data.files.addAll(files?.map((i) => MapEntry(
+    _data.files.addAll(files.map((i) => MapEntry(
         'customfiles',
         MultipartFile.fromFileSync(
           i.path,
           filename: i.path.split(Platform.pathSeparator).last,
           contentType: MediaType.parse('application/json'),
         ))));
-    if (file != null) {
-      _data.files.add(MapEntry(
-          'file',
-          MultipartFile.fromFileSync(file.path,
-              filename: file.path.split(Platform.pathSeparator).last)));
-    }
-    final _result = await _dio.request<String>('/post',
-        queryParameters: queryParameters,
-        options: RequestOptions(
-            method: 'POST',
-            headers: <String, dynamic>{},
-            extra: _extra,
-            baseUrl: baseUrl),
-        data: _data);
-    final value = _result.data;
+    _data.files.add(MapEntry(
+        'file',
+        MultipartFile.fromFileSync(file.path,
+            filename: file.path.split(Platform.pathSeparator).last)));
+    final _result = await _dio.fetch<String>(_setStreamType<String>(
+        Options(method: 'POST', headers: <String, dynamic>{}, extra: _extra)
+            .compose(_dio.options, '/post',
+                queryParameters: queryParameters, data: _data)
+            .copyWith(baseUrl: baseUrl ?? _dio.options.baseUrl)));
+    final value = _result.data!;
     return value;
   }
 
   @override
-  Future<String> postFormData6({files, file}) async {
+  Future<String> postFormData6({required files, required file}) async {
     const _extra = <String, dynamic>{};
     final queryParameters = <String, dynamic>{};
-    queryParameters.removeWhere((k, v) => v == null);
     final _data = FormData();
-    _data.files.addAll(files?.map((i) => MapEntry(
+    _data.files.addAll(files.map((i) => MapEntry(
         'customfiles',
         MultipartFile.fromBytes(
           i,
@@ -536,22 +448,17 @@ class _RestClient implements RestClient {
           file,
           filename: null,
         )));
-    final _result = await _dio.request<String>('/post',
-        queryParameters: queryParameters,
-        options: RequestOptions(
-            method: 'POST',
-            headers: <String, dynamic>{},
-            extra: _extra,
-            baseUrl: baseUrl),
-        data: _data);
-    final value = _result.data;
+    final _result = await _dio.fetch<String>(_setStreamType<String>(
+        Options(method: 'POST', headers: <String, dynamic>{}, extra: _extra)
+            .compose(_dio.options, '/post',
+                queryParameters: queryParameters, data: _data)
+            .copyWith(baseUrl: baseUrl ?? _dio.options.baseUrl)));
+    final value = _result.data!;
     return value;
   }
 
   @override
   Future<String> postFormData4(tasks, file) async {
-    ArgumentError.checkNotNull(tasks, 'tasks');
-    ArgumentError.checkNotNull(file, 'file');
     const _extra = <String, dynamic>{};
     final queryParameters = <String, dynamic>{};
     final _data = FormData();
@@ -560,95 +467,67 @@ class _RestClient implements RestClient {
         'file',
         MultipartFile.fromFileSync(file.path,
             filename: file.path.split(Platform.pathSeparator).last)));
-    final _result = await _dio.request<String>('/post',
-        queryParameters: queryParameters,
-        options: RequestOptions(
-            method: 'POST',
-            headers: <String, dynamic>{},
-            extra: _extra,
-            baseUrl: baseUrl),
-        data: _data);
-    final value = _result.data;
+    final _result = await _dio.fetch<String>(_setStreamType<String>(
+        Options(method: 'POST', headers: <String, dynamic>{}, extra: _extra)
+            .compose(_dio.options, '/post',
+                queryParameters: queryParameters, data: _data)
+            .copyWith(baseUrl: baseUrl ?? _dio.options.baseUrl)));
+    final value = _result.data!;
     return value;
   }
 
   @override
-  Future<String> postFormData5(tasks, map, a, {b, c, d}) async {
-    ArgumentError.checkNotNull(tasks, 'tasks');
-    ArgumentError.checkNotNull(map, 'map');
-    ArgumentError.checkNotNull(a, 'a');
+  Future<String> postFormData5(tasks, map, a,
+      {required b, required c, required d}) async {
     const _extra = <String, dynamic>{};
     final queryParameters = <String, dynamic>{};
-    queryParameters.removeWhere((k, v) => v == null);
     final _data = FormData();
     _data.fields.add(MapEntry('tasks', jsonEncode(tasks)));
     _data.fields.add(MapEntry('map', jsonEncode(map)));
-    if (a != null) {
-      _data.fields.add(MapEntry('a', a.toString()));
-    }
-    if (b != null) {
-      _data.fields.add(MapEntry('b', b.toString()));
-    }
-    if (c != null) {
-      _data.fields.add(MapEntry('c', c.toString()));
-    }
-    if (d != null) {
-      _data.fields.add(MapEntry('d', d));
-    }
-    final _result = await _dio.request<String>('/post',
-        queryParameters: queryParameters,
-        options: RequestOptions(
-            method: 'POST',
-            headers: <String, dynamic>{},
-            extra: _extra,
-            baseUrl: baseUrl),
-        data: _data);
-    final value = _result.data;
+    _data.fields.add(MapEntry('a', a.toString()));
+    _data.fields.add(MapEntry('b', b.toString()));
+    _data.fields.add(MapEntry('c', c.toString()));
+    _data.fields.add(MapEntry('d', d));
+    final _result = await _dio.fetch<String>(_setStreamType<String>(
+        Options(method: 'POST', headers: <String, dynamic>{}, extra: _extra)
+            .compose(_dio.options, '/post',
+                queryParameters: queryParameters, data: _data)
+            .copyWith(baseUrl: baseUrl ?? _dio.options.baseUrl)));
+    final value = _result.data!;
     return value;
   }
 
   @override
   Future<String> queries(queries) async {
-    ArgumentError.checkNotNull(queries, 'queries');
     const _extra = <String, dynamic>{};
     final queryParameters = <String, dynamic>{};
-    queryParameters.addAll(queries ?? <String, dynamic>{});
+    queryParameters.addAll(queries);
     final _data = <String, dynamic>{};
-    final _result = await _dio.request<String>('/demo',
-        queryParameters: queryParameters,
-        options: RequestOptions(
-            method: 'GET',
-            headers: <String, dynamic>{},
-            extra: _extra,
-            baseUrl: baseUrl),
-        data: _data);
-    final value = _result.data;
+    final _result = await _dio.fetch<String>(_setStreamType<String>(
+        Options(method: 'GET', headers: <String, dynamic>{}, extra: _extra)
+            .compose(_dio.options, '/demo',
+                queryParameters: queryParameters, data: _data)
+            .copyWith(baseUrl: baseUrl ?? _dio.options.baseUrl)));
+    final value = _result.data!;
     return value;
   }
 
   @override
   Future<String> queryByEnum(query) async {
-    ArgumentError.checkNotNull(query, 'query');
     const _extra = <String, dynamic>{};
-    final queryParameters = <String, dynamic>{r'tasks': query?.toJson()};
+    final queryParameters = <String, dynamic>{r'tasks': query.toJson()};
     final _data = <String, dynamic>{};
-    final _result = await _dio.request<String>('/enums',
-        queryParameters: queryParameters,
-        options: RequestOptions(
-            method: 'GET',
-            headers: <String, dynamic>{},
-            extra: _extra,
-            baseUrl: baseUrl),
-        data: _data);
-    final value = _result.data;
+    final _result = await _dio.fetch<String>(_setStreamType<String>(
+        Options(method: 'GET', headers: <String, dynamic>{}, extra: _extra)
+            .compose(_dio.options, '/enums',
+                queryParameters: queryParameters, data: _data)
+            .copyWith(baseUrl: baseUrl ?? _dio.options.baseUrl)));
+    final value = _result.data!;
     return value;
   }
 
   @override
   Future<String> namedExample(apiKey, scope, type, {from}) async {
-    ArgumentError.checkNotNull(apiKey, 'apiKey');
-    ArgumentError.checkNotNull(scope, 'scope');
-    ArgumentError.checkNotNull(type, 'type');
     const _extra = <String, dynamic>{};
     final queryParameters = <String, dynamic>{
       r'$apikey': apiKey,
@@ -658,62 +537,58 @@ class _RestClient implements RestClient {
     };
     queryParameters.removeWhere((k, v) => v == null);
     final _data = <String, dynamic>{};
-    final _result = await _dio.request<String>('/get',
-        queryParameters: queryParameters,
-        options: RequestOptions(
-            method: 'GET',
-            headers: <String, dynamic>{},
-            extra: _extra,
-            baseUrl: baseUrl),
-        data: _data);
-    final value = _result.data;
+    final _result = await _dio.fetch<String>(_setStreamType<String>(
+        Options(method: 'GET', headers: <String, dynamic>{}, extra: _extra)
+            .compose(_dio.options, '/get',
+                queryParameters: queryParameters, data: _data)
+            .copyWith(baseUrl: baseUrl ?? _dio.options.baseUrl)));
+    final value = _result.data!;
     return value;
   }
 
   @override
-  Future<String> postFile({file}) async {
+  Future<String> postFile({required file}) async {
     const _extra = <String, dynamic>{};
     final queryParameters = <String, dynamic>{};
-    queryParameters.removeWhere((k, v) => v == null);
     final _data = Stream.fromIterable(file.readAsBytesSync().map((i) => [i]));
-    final _result = await _dio.request<String>('/postfile',
-        queryParameters: queryParameters,
-        options: RequestOptions(
+    final _result = await _dio.fetch<String>(_setStreamType<String>(Options(
             method: 'POST',
             headers: <String, dynamic>{
               r'$Content-Type': 'application/octet-stream',
               r'Ocp-Apim-Subscription-Key': 'abc'
             },
-            extra: _extra,
-            baseUrl: baseUrl),
-        data: _data);
-    final value = _result.data;
+            extra: _extra)
+        .compose(_dio.options, '/postfile',
+            queryParameters: queryParameters, data: _data)
+        .copyWith(baseUrl: baseUrl ?? _dio.options.baseUrl)));
+    final value = _result.data!;
     return value;
   }
 
   @override
   Future<String> testCustomOptions(options) async {
-    ArgumentError.checkNotNull(options, 'options');
     const _extra = <String, dynamic>{};
     final queryParameters = <String, dynamic>{};
     final _data = <String, dynamic>{};
     final newOptions = newRequestOptions(options);
     newOptions.extra.addAll(_extra);
     newOptions.headers.addAll(<String, dynamic>{});
-    final _result = await _dio.request<String>('',
+    final _result = await _dio.fetch<String>(newOptions.copyWith(
+        method: 'GET',
+        baseUrl: baseUrl,
         queryParameters: queryParameters,
-        options: newOptions.merge(method: 'GET', baseUrl: baseUrl),
-        data: _data);
-    final value = _result.data;
+        path: '')
+      ..data = _data);
+    final value = _result.data!;
     return value;
   }
 
-  RequestOptions newRequestOptions(Options options) {
+  RequestOptions newRequestOptions(Options? options) {
     if (options is RequestOptions) {
-      return options;
+      return options as RequestOptions;
     }
     if (options == null) {
-      return RequestOptions();
+      return RequestOptions(path: '');
     }
     return RequestOptions(
       method: options.method,
@@ -729,6 +604,20 @@ class _RestClient implements RestClient {
       maxRedirects: options.maxRedirects,
       requestEncoder: options.requestEncoder,
       responseDecoder: options.responseDecoder,
+      path: '',
     );
+  }
+
+  RequestOptions _setStreamType<T>(RequestOptions requestOptions) {
+    if (T != dynamic &&
+        !(requestOptions.responseType == ResponseType.bytes ||
+            requestOptions.responseType == ResponseType.stream)) {
+      if (T == String) {
+        requestOptions.responseType = ResponseType.plain;
+      } else {
+        requestOptions.responseType = ResponseType.json;
+      }
+    }
+    return requestOptions;
   }
 }